# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/lessons": {
      "get": {
        "operationId": "LessonController_findLessonsForPeriod",
        "parameters": [
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "description": "Start date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "description": "End date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacher_id",
            "required": false,
            "in": "query",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonOutputDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all lessons for a period for current teacher, or if current teacher is admin, get lessons by teacher ID in query or admin's lessons",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/student/{student_id}": {
      "get": {
        "operationId": "LessonController_findLessonsForPeriodAndStudent",
        "parameters": [
          {
            "name": "student_id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "description": "Start date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "description": "End date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLessonsOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all lessons for a period for a student",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/rescheduled": {
      "get": {
        "operationId": "LessonController_findLessonsForReschedule",
        "parameters": [
          {
            "name": "teacher_id",
            "required": false,
            "in": "query",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rescheduled lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonOutputDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all rescheduled lessons for current teacher, or if current teacher is admin, get rescheduled lessons by teacher ID in query or admin's lessons",
        "tags": [
          "Lessons"
        ]
      },
      "post": {
        "operationId": "LessonController_createRescheduledLesson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduledLessonInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rescheduled lesson created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "You are not allowed to reschedule this lesson",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a rescheduled lesson",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/single": {
      "post": {
        "operationId": "LessonController_createSingleLessonByAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SingleLessonInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lesson created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Lesson already booked for this student at this time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a single lesson",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/regular/{student_id}": {
      "post": {
        "operationId": "LessonController_createRegularLesson",
        "parameters": [
          {
            "name": "student_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularLessonsInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Regular lessons created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonOutputDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lesson already booked for this student at this time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create regular lessons for a student",
        "tags": [
          "Lessons"
        ]
      },
      "get": {
        "operationId": "LessonController_getRegularLessons",
        "parameters": [
          {
            "name": "student_id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Regular lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularLessonOutputDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all regular lessons for a student",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/assigned": {
      "get": {
        "operationId": "LessonController_findLessonsByStartDate",
        "parameters": [
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "description": "Start date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lessons have been successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonOutputDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all assigned lessons for a date for current teacher",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/{id}/teacher": {
      "patch": {
        "operationId": "LessonController_changeTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeTeacherDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Teacher changed successfully"
          },
          "400": {
            "description": "Teacher not found or is deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change teacher for a lesson",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/{id}/cancel": {
      "patch": {
        "operationId": "LessonController_cancelLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelLessonDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Lesson cancelled successfully"
          },
          "400": {
            "description": "Lesson not found or is cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lesson not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a lesson",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/{id}/free": {
      "patch": {
        "operationId": "LessonController_manageFreeLessonStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManageFreeLessonStatusDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Free lesson status managed successfully"
          },
          "400": {
            "description": "Lesson not found or is deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manage free lesson status",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/{id}": {
      "delete": {
        "operationId": "LessonController_deleteLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lesson has been successfully deleted"
          },
          "404": {
            "description": "Lesson not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all assigned lessons for a date for current teacher",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/update-lessons-plan-for-period": {
      "patch": {
        "operationId": "LessonController_updateLessonsPlanForPeriod",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Lessons plan update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonsPlanForPeriodDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Lessons plan updated successfully"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update lessons plan for a period",
        "tags": [
          "Lessons"
        ]
      }
    },
...
  },
  "info": {
    "title": "School API",
    "description": "The School API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BadRequestErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "The HTTP status code",
            "example": 400
          },
          "path": {
            "type": "string",
            "description": "The path of the request",
            "example": "/api/v1/users"
          },
          "message": {
            "type": "object",
            "description": "The message of the error",
            "example": [
              "Error message 1",
              "Error message 2"
            ]
          }
        },
        "required": [
          "statusCode",
          "path",
          "message"
        ]
      },
...
      "TeacherBasicOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The id of the teacher",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the teacher",
            "example": "John Doe"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "LessonOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Lesson ID",
            "example": 1
          },
          "student": {
            "description": "Student information",
            "allOf": [
              {
                "$ref": "#/components/schemas/StudentDto"
              }
            ]
          },
          "plan": {
            "description": "Plan information",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDto"
              }
            ]
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Start date with time",
            "example": "2026-06-29T16:07:14.734Z"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "PENDING_PAID"
          },
          "comment": {
            "type": "object",
            "description": "Comment",
            "example": "Comment",
            "nullable": true
          },
          "is_regular": {
            "type": "boolean",
            "description": "Is regular",
            "example": true
          },
          "is_free": {
            "type": "boolean",
            "description": "Is free",
            "example": true
          },
          "is_trial": {
            "type": "boolean",
            "description": "Is trial",
            "example": true
          },
          "rescheduled_lesson_id": {
            "type": "object",
            "description": "Rescheduled lesson ID",
            "example": 1,
            "nullable": true
          },
          "rescheduled_lesson_date": {
            "type": "object",
            "description": "Rescheduled lesson date",
            "example": "2026-06-29T16:07:14.734Z",
            "nullable": true
          },
          "rescheduled_to_lesson_id": {
            "type": "object",
            "description": "Rescheduled to lesson ID",
            "example": 1,
            "nullable": true
          },
          "rescheduled_to_lesson_date": {
            "type": "object",
            "description": "Rescheduled to lesson date",
            "example": "2026-06-29T16:07:14.734Z",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2026-06-29T16:07:14.734Z"
          },
          "teacher": {
            "description": "Teacher information",
            "allOf": [
              {
                "$ref": "#/components/schemas/TeacherBasicOutputDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "student",
          "plan",
          "date",
          "status",
          "comment",
          "is_regular",
          "is_free",
          "is_trial",
          "rescheduled_lesson_id",
          "rescheduled_lesson_date",
          "rescheduled_to_lesson_id",
          "rescheduled_to_lesson_date",
          "created_at",
          "teacher"
        ]
      },
      "StudentLessonsOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Student name",
            "example": "John Doe"
          },
          "class": {
            "type": "number",
            "description": "Student class",
            "example": 1
          },
          "canceled_lessons": {
            "type": "number",
            "description": "Canceled lessons",
            "example": 1
          },
          "missed_lessons": {
            "type": "number",
            "description": "Missed lessons",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "class",
          "canceled_lessons",
          "missed_lessons"
        ]
      },
      "SingleLessonInputDto": {
        "type": "object",
        "properties": {
          "plan_id": {
            "type": "number",
            "description": "Plan ID",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Start date",
            "example": "2026-06-29T16:07:14.728Z"
          },
          "isFree": {
            "type": "boolean",
            "description": "Is free",
            "example": true
          },
          "isTrial": {
            "type": "boolean",
            "description": "Is trial",
            "example": true
          }
        },
        "required": [
          "plan_id",
          "student_id",
          "teacher_id",
          "start_date",
          "isFree",
          "isTrial"
        ]
      },
      "RescheduledLessonInputDto": {
        "type": "object",
        "properties": {
          "rescheduled_lesson_id": {
            "type": "number",
            "description": "Rescheduled lesson ID",
            "example": 1
          },
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Rescheduled lesson date",
            "example": "2026-06-29T16:07:14.740Z"
          }
        },
        "required": [
          "rescheduled_lesson_id",
          "start_date"
        ]
      },
      "RegularLessonsInputDto": {
        "type": "object",
        "properties": {}
      },
      "RegularLessonOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Regular lesson ID",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          },
          "teacher_id": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          },
          "start_time": {
            "type": "string",
            "description": "Start time",
            "example": "10:00"
          },
          "week_day": {
            "type": "string",
            "description": "Week day",
            "example": "MONDAY"
          },
          "start_period_date": {
            "format": "date-time",
            "type": "string",
            "description": "Start period date",
            "example": "2026-06-29T16:07:14.755Z"
          },
          "end_period_date": {
            "format": "date-time",
            "type": "string",
            "description": "End period date",
            "example": "2026-06-29T16:07:14.755Z"
          },
          "plan": {
            "description": "Plan",
            "example": {
              "id": 1,
              "plan_name": "10 minutes",
              "plan_price": 1000000,
              "plan_currency": "USD",
              "duration": 10,
              "plan_type": "INDIVIDUAL",
              "deleted_at": null,
              "created_at": "2026-06-29T16:07:14.755Z"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/PlanDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "student_id",
          "teacher_id",
          "start_time",
          "week_day",
          "start_period_date",
          "end_period_date",
          "plan"
        ]
      },
      "ChangeTeacherDto": {
        "type": "object",
        "properties": {
          "teacher_id": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          }
        },
        "required": [
          "teacher_id"
        ]
      },
      "CancelLessonDto": {
        "type": "object",
        "properties": {}
      },
      "ManageFreeLessonStatusDto": {
        "type": "object",
        "properties": {
          "isFree": {
            "type": "boolean",
            "description": "Is free",
            "example": true
          }
        },
        "required": [
          "isFree"
        ]
      },
      "UpdateLessonsPlanForPeriodDto": {
        "type": "object",
        "properties": {
          "start_date": {
            "format": "date-time",
            "type": "string",
            "description": "Start date",
            "example": "2026-06-29T16:07:14.756Z"
          },
          "end_date": {
            "format": "date-time",
            "type": "string",
            "description": "End date",
            "example": "2026-06-29T16:07:14.756Z"
          },
          "new_plan_id": {
            "type": "number",
            "description": "Plan ID",
            "example": 1
          },
          "old_plan_id": {
            "type": "number",
            "description": "Old plan ID",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "description": "Student ID",
            "example": 1
          }
        },
        "required": [
          "start_date",
          "end_date",
          "new_plan_id",
          "old_plan_id",
          "student_id"
        ]
      },
...
    }
  }
}